geo_agent/requirements.txt
pandas
geopandas
scikit-learn
google-cloud-aiplatform
google-generativeai
pydantic
tenacity
ipython
fsspec
gcsfs